+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 0.0.1
-
-* TODO: Describe initial release.
+- Added `FlutterReverbOptions` class.
+- Added `FlutterReverb` class.
+## 0.0.2
+- Changed `listen` method to accept `isPrivate` parameter.
+- Changed `authenticate` method to be private.
+- Changed `subscribe` method to be private.
+- Removed broken tests.
+- TODO: Rewrite tests.